src-main/requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
parsedatetime
pyenchant
asyncpg
python-dateutil
pillow
toml